Fix regression in mount.cifs with guest mount option
mount.cifs was erroneously passing the empty password and password2
strings to the linux smb client for "guest" mounts. Handling empty
password2 has not been implemented on the client yet, which ends up
failing the mount with "cifs: Bad value for password2".

This patch contains fixes for the mentioned scenario.

Signed-off-by: Meetakshi Setiya <msetiya@microsoft.com>